Rename PersistentActor to PersistentBehavior, #24071
