+++ akka-persistence-typed/src/test/scala/akka/persistence/typed/scaladsl/PersistentBehaviorCompileOnlyTest.scala
@@ -10,9 +10,9 @@ import akka.actor.typed.scaladsl.ActorBehavior
 import akka.actor.typed.scaladsl.ActorContext
 import akka.actor.typed.scaladsl.TimerScheduler
 
-object PersistentActorCompileOnlyTest {
+object PersistentBehaviorCompileOnlyTest {
 
-  import akka.persistence.typed.scaladsl.PersistentActor._
+  import akka.persistence.typed.scaladsl.PersistentBehavior._
 
   object Simple {
     //#command
@@ -42,7 +42,7 @@ object PersistentActorCompileOnlyTest {
 
     //#behavior
     val simpleBehavior: PersistentBehavior[SimpleCommand, SimpleEvent, ExampleState] =
-      PersistentActor.immutable[SimpleCommand, SimpleEvent, ExampleState](
+      PersistentBehavior.immutable[SimpleCommand, SimpleEvent, ExampleState](
         persistenceId = "sample-id-1",
         initialState = ExampleState(Nil),
         commandHandler = commandHandler,
@@ -62,7 +62,7 @@ object PersistentActorCompileOnlyTest {
 
     case class ExampleState(events: List[String] = Nil)
 
-    PersistentActor.immutable[MyCommand, MyEvent, ExampleState](
+    PersistentBehavior.immutable[MyCommand, MyEvent, ExampleState](
       persistenceId = "sample-id-1",
 
       initialState = ExampleState(Nil),
@@ -106,7 +106,7 @@ object PersistentActorCompileOnlyTest {
         .foreach(sender ! _)
     }
 
-    PersistentActor.immutable[Command, Event, EventsInFlight](
+    PersistentBehavior.immutable[Command, Event, EventsInFlight](
       persistenceId = "recovery-complete-id",
 
       initialState = EventsInFlight(0, Map.empty),
@@ -149,7 +149,7 @@ object PersistentActorCompileOnlyTest {
     sealed trait Event
     case class MoodChanged(to: Mood) extends Event
 
-    val b: Behavior[Command] = PersistentActor.immutable[Command, Event, Mood](
+    val b: Behavior[Command] = PersistentBehavior.immutable[Command, Event, Mood](
       persistenceId = "myPersistenceId",
       initialState = Happy,
       commandHandler = CommandHandler.byState {
@@ -190,7 +190,7 @@ object PersistentActorCompileOnlyTest {
 
     case class State(tasksInFlight: List[Task])
 
-    PersistentActor.immutable[Command, Event, State](
+    PersistentBehavior.immutable[Command, Event, State](
       persistenceId = "asdf",
       initialState = State(Nil),
       commandHandler = CommandHandler.command {
@@ -217,7 +217,7 @@ object PersistentActorCompileOnlyTest {
 
     def worker(task: Task): Behavior[Nothing] = ???
 
-    PersistentActor.immutable[Command, Event, State](
+    PersistentBehavior.immutable[Command, Event, State](
       persistenceId = "asdf",
       initialState = State(Nil),
       commandHandler = (ctx, _, cmd) ⇒ cmd match {
@@ -280,7 +280,7 @@ object PersistentActorCompileOnlyTest {
           .persist[Event, List[Id]](ItemAdded(id))
           .andThen(metadataRegistry ! GetMetaData(id, adapt))
 
-      PersistentActor.immutable[Command, Event, List[Id]](
+      PersistentBehavior.immutable[Command, Event, List[Id]](
         persistenceId = "basket-1",
         initialState = Nil,
         commandHandler =
@@ -342,7 +342,7 @@ object PersistentActorCompileOnlyTest {
       if (currentState == newMood) Effect.none
       else Effect.persist(MoodChanged(newMood))
 
-    PersistentActor.immutable[Command, Event, Mood](
+    PersistentBehavior.immutable[Command, Event, Mood](
       persistenceId = "myPersistenceId",
       initialState = Sad,
       commandHandler = (_, state, cmd) ⇒
@@ -378,7 +378,7 @@ object PersistentActorCompileOnlyTest {
 
     class State
 
-    PersistentActor.immutable[Command, Event, State](
+    PersistentBehavior.immutable[Command, Event, State](
       persistenceId = "myPersistenceId",
       initialState = new State,
       commandHandler = CommandHandler.command {
