+++ akka-persistence-typed/src/test/scala/docs/akka/persistence/typed/InDepthPersistentBehaviorSpec.scala
@@ -5,10 +5,10 @@ package docs.akka.persistence.typed
 
 import akka.Done
 import akka.actor.typed.{ ActorRef, Behavior }
-import akka.persistence.typed.scaladsl.PersistentActor
-import akka.persistence.typed.scaladsl.PersistentActor.{ CommandHandler, Effect }
+import akka.persistence.typed.scaladsl.PersistentBehavior
+import akka.persistence.typed.scaladsl.PersistentBehavior.{ CommandHandler, Effect }
 
-object InDepthPersistentActorSpec {
+object InDepthPersistentBehaviorSpec {
 
   //#event
   sealed trait BlogEvent extends Serializable
@@ -117,7 +117,7 @@ object InDepthPersistentActorSpec {
 
   //#behavior
   def behavior: Behavior[BlogCommand] =
-    PersistentActor.immutable[BlogCommand, BlogEvent, BlogState](
+    PersistentBehavior.immutable[BlogCommand, BlogEvent, BlogState](
       persistenceId = "abc",
       initialState = BlogState.empty,
       commandHandler,
@@ -125,6 +125,6 @@ object InDepthPersistentActorSpec {
   //#behavior
 }
 
-class InDepthPersistentActorSpec {
+class InDepthPersistentBehaviorSpec {
 
 }
